% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cleaning_log.R
\name{create_cleaning_log}
\alias{create_cleaning_log}
\title{Bind tables containing cleaning issues}
\usage{
create_cleaning_log(spotted_by = "BGD_Data_Unit", ...)
}
\arguments{
\item{spotted_by}{Name of the person/unit who identified the issue. Default vale is "BGD_Data_Unit"}

\item{...}{Data sets which should be added in the cleaning log}
}
\description{
Bind tables containing cleaning issues
}
