% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixing data type.R
\name{fix_data_type}
\alias{fix_data_type}
\title{Fix data type from the dataset}
\usage{
fix_data_type(
  df,
  remove_all_NA_col = T,
  na.string = c("", "NA"),
  character_cols = NULL
)
}
\arguments{
\item{df}{The data set}

\item{remove_all_NA_col}{A logical parameter. Ture will remove the columns where all values are NA}

\item{kobo_survey_sheet}{The defult (NUll) will create additional table for multiple choice questions}

\item{remove_multiple_choice_concat}{A logical parameter. Ture will remove the aggregated columns (for multiple choice question) from the dataset}
}
\description{
Fix data type from the dataset
}
