% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_multiple_choice_from_data.R
\name{fix_data_type_frm_kobo}
\alias{fix_data_type_frm_kobo}
\title{Generate dataset for butteR::survey_collaps()}
\usage{
fix_data_type_frm_kobo(
  df,
  kobo_servey_sheet,
  remove_multiple_choice_concat = F,
  remove_all_NA_col = T
)
}
\arguments{
\item{df}{The data set}

\item{kobo_servey_sheet}{The defult (NUll) will create additional table for multiple choice questions}

\item{remove_multiple_choice_concat}{A logical parameter. Ture will remove the aggregated columns (for multiple choice question) from the dataset}

\item{remove_all_NA_col}{A logical parameter. Ture will remove the columns where all values are NA}
}
\description{
Generate dataset for butteR::survey_collaps()
}
