% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_na_response.R
\name{get_na_response_rates}
\alias{get_na_response_rates}
\title{Credit goes to Zack Arno
Get NA percentages and frequency for each column in the data set}
\usage{
get_na_response_rates(data)
}
\arguments{
\item{data}{The dataset (data.frame).}
}
\value{
frequency and proportion table of NA rates.
}
\description{
Credit goes to Zack Arno
Get NA percentages and frequency for each column in the data set
}
