% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml_to_shp.R
\name{kml_to_shp}
\alias{kml_to_shp}
\title{Use to convert KMLs to shapefiles}
\usage{
kml_to_shp(input, output)
}
\arguments{
\item{input}{Input folder from where KMLs will be converted to shapefile.}

\item{output}{Output folder in where converted KMLs will be saved.}
}
\description{
Use to convert KMLs to shapefiles
}
