% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_checks.R
\name{logical_check}
\alias{logical_check}
\title{List of logical issues in the data set}
\usage{
logical_check(
  df,
  uuid,
  logic_list_df,
  logic_col_name,
  add_description = T,
  col_name_issue
)
}
\arguments{
\item{df}{Data frame}

\item{uuid}{UUID column name in the data frame.}

\item{logic_list_df}{Data frame with a list of checks. This must include a column filled with all logic for the checks.}

\item{logic_col_name}{Name of the column of logic_list_df data frame where all the logic are listed}

\item{add_description}{TRUE [default ], if you want to add a description for each logic.}

\item{col_name_issue}{Name of the column of logic_list_df data frame where the description for each logic is listed}
}
\value{
A data frame where all the issues will be listed as a cleaning log format
}
\description{
List of logical issues in the data set
}
