% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pic_to_pdf.R
\name{pic_to_pdf}
\alias{pic_to_pdf}
\title{Use to convert images (.jpeg,.jpg,.png) to pdfs}
\usage{
pic_to_pdf(input, output)
}
\arguments{
\item{input}{Input folder from where picture will be converted to pdfs.}

\item{output}{Output folder in where converted pdfs will be saved.}
}
\description{
Use to convert images (.jpeg,.jpg,.png) to pdfs
}
