% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_survey.R
\name{quick_survey}
\alias{quick_survey}
\title{Quick Surveys}
\usage{
quick_survey(
  df,
  consent_column_name,
  audit_node,
  uuid_col_name = "X_uuid",
  start_question_node,
  end_question_node,
  min_allowable_duration,
  audit_yes
)
}
\arguments{
\item{df}{data set}

\item{consent_column_name}{column where surevy consent status is stored (default is X_uuid)}

\item{uuid_col_name}{column contining unique identifier}

\item{start_question_node}{start question node from where calculation for time duration will start (audit node is excluded)}

\item{end_question_node}{end question node to where calculation for time duration will end (audit node is excluded)}

\item{min_allowable_duration}{Minumum allowable duration}

\item{audit_yes}{output from butteR::load_audit()}
}
\value{
Quick survey list
}
\description{
Quick Surveys
}
