% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sheets_as_csv_format.R
\name{read_all_sheet_as_csv_format}
\alias{read_all_sheet_as_csv_format}
\title{read all the sheets from a Excel file}
\usage{
read_all_sheet_as_csv_format(dataset_path)
}
\arguments{
\item{dataset_path}{location of the dataset}

\item{kobo_tool_path}{location of the kobo tool}
}
\value{
All of the sheet as individual dataframe
}
\description{
read all the sheets from a Excel file
}
