% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculate_concerted_cols.R
\name{recalculate_concerted_col_for_select_multiple}
\alias{recalculate_concerted_col_for_select_multiple}
\title{This function recreates the concerted columns for select multiple questions}
\usage{
recalculate_concerted_col_for_select_multiple(df, uuid = "X_uuid")
}
\arguments{
\item{df}{data frame}

\item{uuid}{Unique ID column(Default is X_uuid).}
}
\description{
This function recreates the concerted columns for select multiple questions
}
