% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_chi_square_test.R
\name{simple_chi_square_test}
\alias{simple_chi_square_test}
\title{Simple pair wise chi-square test}
\usage{
simple_chi_square_test(
  survery_object,
  dap,
  independent_variable_col_in_dap,
  dependent_variable_col_in_dap
)
}
\arguments{
\item{survery_object}{Survey object from srvyr or survey package}

\item{dap}{a data frame providing the names of dependent and independent variable}

\item{independent_variable_col_in_dap}{column name that specify the column of independent variable in the the dap}

\item{dependent_variable_col_in_dap}{column name that specify the column of dependent variable in the the dap}
}
\description{
The count is only based on independent variable, not a cross table of dependent and independent variable
All the independent variable must be a binary variable to run the function
}
