% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_test.R
\name{statistical_test}
\alias{statistical_test}
\title{Function for one way ANOVA/T/Chi-square test}
\usage{
statistical_test(
  data,
  kobo_survey_sheet,
  variable_1,
  target_Variables,
  ci = 0.95,
  simulate.p.value = T
)
}
\arguments{
\item{data}{Data}

\item{kobo_survey_sheet}{Questions/survey sheet from the kobo tool. It should be a data frame.}

\item{variable_1}{Mostly independent variable (It might be the dis/aggregation level). It should be a character vector.}

\item{target_Variables}{Mostly dependent variables (On which variables you want to apply the statistical test)}

\item{ci}{Confidence interval. The default is .95}

\item{simulate.p.value}{A logical vector indicating whether to compute p-values by Monte Carlo simulation. This parameter only applied to Chi square test. TRUE is default. FLASE is preferable when you have large data set.}
}
\description{
Function for one way ANOVA/T/Chi-square test
}
\details{
Matrix for Statistical test (when to use what test)
\tabular{rrrrr}{
  \strong{} \tab \strong{Categorical} \tab \strong{Numerical}  \cr
  \strong{Categorical} \tab Chi Square Test \tab T test: When there is only 2 groups/ ANOVA: When there is more than 2 groups \cr
  \strong{Numerical} \tab Regression (Not included in this function) \tab Perason's Correlation test
}
}
