% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_duration.R
\name{survey_duration_from_audit}
\alias{survey_duration_from_audit}
\title{Quick Surveys}
\usage{
survey_duration_from_audit(
  data,
  audit_zip_path,
  filter = T,
  filter_column = "informed_consent",
  filter_choice = "yes",
  uuid_col_name = "X_uuid",
  start_question,
  end_question
)
}
\arguments{
\item{data}{Data set}

\item{audit_zip_path}{Location for audit zip file}

\item{filter}{Yes, if you want to apply}

\item{filter_column}{Name of filter column}

\item{filter_choice}{Filter choice name}

\item{uuid_col_name}{Column containing unique identifier}

\item{start_question}{Start question, where calculation for time duration will start.}

\item{end_question}{End question node to where calculation for time duration will end.}
}
\value{
Survey duration
}
\description{
Quick Surveys
}
