% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suvrey_frequency_by_choices.R
\name{survey_frequency_by_choices}
\alias{survey_frequency_by_choices}
\title{Generate Frequency Table For Each Choices}
\usage{
survey_frequency_by_choices(
  df,
  aggregation_level = NULL,
  variables_to_analyze,
  kobo_servey_sheet = NULL
)
}
\arguments{
\item{df}{The data set}

\item{aggregation_level}{Aggregation level(s) to analyze data by. The default (NULL) will fully aggregate data to provide one value per variable. Argument also accepts vecors to aggregate upto two variables.}

\item{variables_to_analyze}{Vector containing all variable names to analyze and include in summary table.}

\item{kobo_servey_sheet}{Output folder in where converted KMLs will be saved.The defult (NUll) will create additional table for multiple choice questions}
}
\description{
Generate Frequency Table For Each Choices
}
