% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suvrey_frequency_by_questions.R
\name{survey_frequency_by_questions}
\alias{survey_frequency_by_questions}
\title{Generate Frequency Table For Questions}
\usage{
survey_frequency_by_questions(df, aggregation_level = NULL)
}
\arguments{
\item{df}{The data set}

\item{aggregation_level}{Aggregation level(s) to analyze data by. The default (NULL) will fully aggregate data to provide one value per variable. Argument also accepts vecors to aggregate upto two variables.}
}
\description{
Generate Frequency Table For Questions
}
