% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted statistical test.R
\name{weighted_pearson_test}
\alias{weighted_pearson_test}
\title{Weighted pearsons test}
\usage{
weighted_pearson_test(data, dep_var, ind_var, strata, survey_weights)
}
\arguments{
\item{data}{dataset}

\item{dep_var}{must be numeric and only one value can be use.}

\item{ind_var}{must be numeric, multiple values are allowed}

\item{strata}{Column name in the dataset for strata}

\item{survey_weights}{Column name in the dataset for survey weights}
}
\value{
outcome of left join based on the primary key
}
\description{
Weighted pearsons test
}
