% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted statistical test.R
\name{weighted_t_test}
\alias{weighted_t_test}
\title{Weighted T test}
\usage{
weighted_t_test(
  data,
  binary_variable,
  non_binary_variable,
  strata,
  survey_weights
)
}
\arguments{
\item{data}{dataset}

\item{binary_variable}{must be binary variable and only one value can be use.}

\item{non_binary_variable}{must be numeric, multiple values are allowed.}

\item{strata}{Column name in the dataset for strata}

\item{survey_weights}{Column name in the dataset for survey weights}
}
\value{
outcome of left join based on the primary key
}
\description{
Weighted T test
}
