% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_excel_as_reach_format.R
\name{write_excel_as_reach_format}
\alias{write_excel_as_reach_format}
\title{This function provide excel file as REACH format}
\usage{
write_excel_as_reach_format(write_list, output_path, cols_for_color = NULL)
}
\arguments{
\item{write_list}{a list file (can be compile with single or multiple dataframe)}

\item{output_path}{Path for output file}

\item{cols_for_color}{Column name in the dataframe which should be use for colorizing the cell. The default is null.}
}
\value{
Nicely formatted excel file
}
\description{
This function provide excel file as REACH format
}
