% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_multiple_choice_from_data.R
\name{fix_data_type_frm_kobo}
\alias{fix_data_type_frm_kobo}
\title{Generate dataset for butteR::survey_collaps()}
\usage{
fix_data_type_frm_kobo(df, kobo_servey_sheet)
}
\arguments{
\item{df}{The data set}

\item{kobo_servey_sheet}{The defult (NUll) will create additional table for multiple choice questions}
}
\description{
Generate dataset for butteR::survey_collaps()
}
